+++ Modules/Tenancy/routes/web.php
@@ -2,6 +2,7 @@
 
 /**
  * @author  MEMORA solutions <info@memora.ca> (https://memora.solutions)
+ *
  * @project memora/laravel-saas-boilerplate
  */
 
@@ -10,15 +11,15 @@ declare(strict_types=1);
 use Illuminate\Support\Facades\Route;
 use Modules\Tenancy\Http\Controllers\Admin\TenantController;
 
-Route::middleware(['web', 'auth', 'permission:manage_tenants'])
+Route::middleware(['web', 'auth'])
     ->prefix('admin/tenants')
     ->name('admin.tenants.')
     ->group(function () {
-        Route::get('/', [TenantController::class, 'index'])->name('index');
-        Route::get('/create', [TenantController::class, 'create'])->name('create');
-        Route::post('/', [TenantController::class, 'store'])->name('store');
-        Route::get('/{tenant}', [TenantController::class, 'show'])->name('show');
-        Route::get('/{tenant}/edit', [TenantController::class, 'edit'])->name('edit');
-        Route::put('/{tenant}', [TenantController::class, 'update'])->name('update');
-        Route::delete('/{tenant}', [TenantController::class, 'destroy'])->name('destroy');
+        Route::get('/', [TenantController::class, 'index'])->name('index')->middleware('permission:view_tenants');
+        Route::get('/create', [TenantController::class, 'create'])->name('create')->middleware('permission:create_tenants');
+        Route::get('/{tenant}', [TenantController::class, 'show'])->name('show')->middleware('permission:view_tenants');
+        Route::post('/', [TenantController::class, 'store'])->name('store')->middleware('permission:create_tenants');
+        Route::get('/{tenant}/edit', [TenantController::class, 'edit'])->name('edit')->middleware('permission:update_tenants');
+        Route::put('/{tenant}', [TenantController::class, 'update'])->name('update')->middleware('permission:update_tenants');
+        Route::delete('/{tenant}', [TenantController::class, 'destroy'])->name('destroy')->middleware('permission:delete_tenants');
     });
